fix: Resolve PHPStan and code style violations
- Split ReferencePropertyFilter and ReferenceIdFilter into separate files (PSR-12 compliance)
- Remove calls to non-existent methods (greaterThanEqual, lessThanEqual, containsAll)
- Add proper type annotations for array parameters
- Add imports for separated classes in ReferenceFilter
- Ensure each class is in its own file following PSR-12 standards
